Перенесла всё на новый activity. Добавила переход на dbactivity при долгом нажатии на кнопку "назад".
